Add files via upload
Standalone Event Management Java based software which utilizes MySQL for Data Manipulation.
It is used to Manage and Manipulate Data of Events, Participants and Organizing Members.
Implemented techniques to overcome the Static Dimensions of
Java Frames and Objects. Every Frame and Object will dynamically adjust according to the Screen Resolution.